Splitting up the MagnetClasses.cc file, adding new magnet type TARGET, allowing --object as well as --magnet
