+++ include/MagnetClasses.hh
@@ -120,8 +120,21 @@ private:
     G4String absorberMaterialName;
     G4Material* absorberMaterial = NULL;
     G4double width  = 10.0*mm;  //[G4 length units]
-    G4double height = 50*mm;    //[G4 length units]
-    G4double radius = 50*mm;    //[G4 length units]
+    G4double height = 50.0*mm;  //[G4 length units]
+    G4double radius = 50.0*mm;  //[G4 length units]
+};
+
+class MagnetTARGET : public MagnetBase {
+public:
+    MagnetTARGET(G4double zPos_in, G4bool doRelPos_in, G4double length_in, G4double gradient_in,
+                      std::map<G4String,G4String> &keyValPairs_in, DetectorConstruction* detCon_in,
+                      G4String magnetName_in);
+    virtual void Construct();
+private:
+    G4String targetMaterialName;
+    G4Material* targetMaterial = NULL;
+    G4double width  = 10.0*mm;  //[G4 length units]
+    G4double height = 10.0*mm;  //[G4 length units]
 };
 
 #endif
